@@ -1,13 +1,13 @@
 {
-  "name": "el-form-auto",
-  "version": "1.0.11",
-  "description": "Dynamic Load Element Form for Vue.js",
-  "main": "lib/ELFormAuto.umd.js",
+  "name": "lea-auto-form",
+  "version": "3.1.18",
+  "description": "Dynamic Load Element Form for Vue3",
+  "main": "lib/index.umd.js",
   "files": [
-    "lib"
+    "src"
   ],
   "scripts": {
-    "build-lib": "vue-cli-service build --target lib --name ElFormAuto 'src/el-form-auto/Form.vue'",
+    "build-lib": "vue-cli-service build --target lib --name LeaAutoForm 'src/lea-auto-form/Form.vue'",
     "lint": "vue-cli-service lint",
     "demo": "vue-cli-service serve",
     "build-demo": "vue-cli-service build",
@@ -30,13 +30,13 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "element-ui": "^2.4.5",
-    "vue": "^2.6.14"
+    "vue": "^3.2.45",
+    "element-plus": "2.3.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.6.0",
+    "@vue/cli-plugin-babel": "^4.5.0",
     "@vue/cli-plugin-eslint": "^3.6.0",
-    "@vue/cli-service": "^3.6.0",
+    "@vue/cli-service": "^4.5.0",
     "@vue/eslint-config-standard": "^4.0.0",
     "babel-eslint": "^10.0.1",
     "babel-plugin-component": "^1.1.1",
@@ -45,7 +45,7 @@
     "sass": "^1.18.0",
     "sass-loader": "^7.1.0",
     "vue-cli-plugin-element": "^1.0.1",
-    "vue-template-compiler": "^2.5.21"
+    "@vue/compiler-sfc": "^3.1.1"
   },
   "eslintConfig": {
     "root": true,
